MV: docker adapter: Move some bits to adapters.utils
These parts will be useful for the upcoming skopeo adapter as well.
